Update tests, setup, gitignore
